updated IDE, changed floats to doubles, lots of refactoring of components and such

+++ src/com/kadware/koala/modules/FixedMixerModule.java
@@ -104,12 +104,12 @@ public class FixedMixerModule extends Module {
         _panMod3.advance();
         _panMod4.advance();
 
-        float left = _pan1LeftOut.getCurrentValue()
+        double left = _pan1LeftOut.getCurrentValue()
             + _pan2LeftOut.getCurrentValue()
             + _pan3LeftOut.getCurrentValue()
             + _pan4LeftOut.getCurrentValue();
 
-        float right = _pan1RightOut.getCurrentValue()
+        double right = _pan1RightOut.getCurrentValue()
             + _pan2RightOut.getCurrentValue()
             + _pan3RightOut.getCurrentValue()
             + _pan4RightOut.getCurrentValue();
